Move state calculations to a new abstract class
