+++ cirq/__init__.py
@@ -200,6 +200,7 @@ from cirq.sim import (
     StepResult,
     SimulatesFinalWaveFunction,
     SimulatesIntermediateWaveFunction,
+    State,
     bloch_vector_from_state_vector,
     density_matrix_from_state_vector,
     dirac_notation,
